Capstone/employeeManagement/src/main/java/com/employee/employeeManagement/dto/RequestResourceDto.java
package com.employee.employeeManagement.dto;


/**
 * Data Transfer Object (DTO) representing a request resource.
 * This class is used for transferring data
 * related to employee management requests.
 */
public class RequestResourceDto {

    /**
     * The comment associated with the request.
     */
    private String comment;

    /**
     * The ID of the manager responsible for the request.
     */
    private Long managerId;

    /**
     * The ID of the employee associated with the request.
     */
    private Long employeeId;

    /**
     * The ID of the project related to the request.
     */
    private Long projectId;

    /**
     * Get the comment associated with the request.
     *
     * @return The comment as a string.
     */
    public String getComment() {
        return comment;
    }

    /**
     * Set the comment associated with the request.
     *
     * @param commentParam The comment to set.
     */
    public void setComment(final String commentParam) {
        this.comment = commentParam;
    }

    /**
     * Get the ID of the manager responsible for the request.
     *
     * @return The manager's ID.
     */
    public Long getManagerId() {
        return managerId;
    }

    /**
     * Set the ID of the manager responsible for the request.
     *
     * @param managerIdParam The manager's ID to set.
     */
    public void setManagerId(final Long managerIdParam) {
        this.managerId = managerIdParam;
    }

    /**
     * Get the ID of the employee associated with the request.
     *
     * @return The employee's ID.
     */
    public Long getEmployeeId() {
        return employeeId;
    }

    /**
     * Set the ID of the employee associated with the request.
     *
     * @param employeeIdParam The employee's ID to set.
     */
    public void setEmployeeId(final Long employeeIdParam) {
        this.employeeId = employeeIdParam;
    }

    /**
     * Get the ID of the project related to the request.
     *
     * @return The project's ID.
     */
    public Long getProjectId() {
        return projectId;
    }

    /**
     * Set the ID of the project related to the request.
     *
     * @param projectIdParam The project's ID to set.
     */
    public void setProjectId(final Long projectIdParam) {
        this.projectId = projectIdParam;
    }
}
